add gameimg endpoint
